Fixing bankdata document field to bigint

+++ database/migrations/2017_12_18_125246_create_nametable_table.php
@@ -48,7 +48,7 @@ class CreateNametableTable extends Migration
             $table->integer('userid')->unsigned();
             $table->integer('bankid');
             $table->string('fullname');
-            $table->integer('document');
+            $table->bigInteger('document');
             $table->string('agency');
             $table->string('account');
             $table->integer('type');
